+++ Source/GeneDefOf.cs
@@ -140,6 +140,7 @@ namespace RJW_Genes
         public static readonly GeneDef rjw_genes_infectious_homosexuality;
 		public static readonly GeneDef rjw_genes_infectious_hypersexuality;
         public static readonly GeneDef rjw_genes_stretcher;
+		public static readonly GeneDef rjw_genes_infectious_blocked_masturbation;
 
         //Other Defs
         public static readonly XenotypeDef rjw_genes_succubus;